Fix: Change postId to int postId was being checked as string which threw an error. The ids are stored as integer

+++ controllers/postController.js
@@ -15,7 +15,7 @@ async function getPost(req, res) {
     const { postId } = req.params;
 
     const post = await prisma.post.findFirst({
-      where: { id: parseFloat(postId) },
+      where: { id: parseInt(postId) },
       include: {
         author: {
           select: {
